프레임 계산이 잘못됫었음

+++ Macros.h
@@ -1,4 +1,4 @@
 #pragma once
 
 #define GET_LAST_ERROR printf("%s::%d %d", __FILE__, __LINE__, GetLastError())
-#define MS_PER_UPDATE (unsigned long)1.0 / 60.0
+#define MS_PER_UPDATE (double)1.0 / 60.0
